changed inode id from 32bit to 64bit

+++ toolchain/udbfslib/allocate_inode_id.c
@@ -1,4 +1,4 @@
-// $Header: /home/xubuntu/berlios_backup/github/tmp-cvs/uuu/Repository/toolchain/udbfslib/allocate_inode_id.c,v 1.1 2003/10/11 13:14:19 bitglue Exp $
+// $Header: /home/xubuntu/berlios_backup/github/tmp-cvs/uuu/Repository/toolchain/udbfslib/allocate_inode_id.c,v 1.2 2003/10/12 18:13:23 instinc Exp $
 
 #define _LARGEFILE_SOURCE
 #define _LARGEFILE64_SOURCE
@@ -11,10 +11,10 @@
 
 
 
-uint32_t	udbfs_allocate_inode_id(
+uint64_t	udbfs_allocate_inode_id(
     UDBFSLIB_MOUNT	*mount ) {
 
-  return( (uint32_t)udbfslib_allocate_bit(
+  return( udbfslib_allocate_bit(
 	mount->inode_bitmap,
 	mount->inode_bitmap_size,
 	&mount->free_inode_count) );
