+++ toolchain/udbfslib/free_inode.c
@@ -1,4 +1,4 @@
-// $Header: /home/xubuntu/berlios_backup/github/tmp-cvs/uuu/Repository/toolchain/udbfslib/free_inode.c,v 1.1 2003/10/11 13:14:19 bitglue Exp $
+// $Header: /home/xubuntu/berlios_backup/github/tmp-cvs/uuu/Repository/toolchain/udbfslib/free_inode.c,v 1.2 2003/10/12 18:10:39 instinc Exp $
 
 #define _LARGEFILE_SOURCE
 #define _LARGEFILE64_SOURCE
@@ -22,7 +22,7 @@ static int		udbfslib_deallocate_bit(
 
 int		udbfs_free_inode(
     UDBFSLIB_MOUNT	*mount,
-    uint32_t		inode_id ) {
+    uint64_t		inode_id ) {
 
   UDBFSLIB_INODE	*inode;
